클라이언트에서 플레이어 이동 확인

+++ source/Client/Client/Player.h
@@ -50,7 +50,7 @@ public:
 	CCamera* GetCamera() { return(m_pCamera); }
 	void SetCamera(CCamera* pCamera) { m_pCamera = pCamera; }
 
-	void Move(ULONG nDirection, float fDistance, bool bVelocity = false);
+	void Move(DWORD nDirection, float fDistance, bool bVelocity = false);
 	void Move(const XMFLOAT3& xmf3Shift, bool bVelocity = false);
 	void Move(float fxOffset = 0.0f, float fyOffset = 0.0f, float fzOffset = 0.0f);
 	void ResetPlayerPos() { m_xmf3Position = XMFLOAT3(0.0f, 0.0f, 0.0f); }
